# Use snprintf in ToStringPrecisionLocale()