Replacement of `std::ostringstream` with `std::snprintf` in function `ToStringPrecisionLocale()`, as proposed by @biodranik in [PR 8896](https://github.com/organicmaps/organicmaps/pull/8896), in order to obtain a better performance.

Columns:
* Old: `std::ostringstream`
* New: `std::snprintf`
* Ratio: New / Old * 100

Sample of values:
```
Old: 61269 ns, New:  8961 ns (Ratio 15 %)
Old: 65807 ns, New: 11539 ns (Ratio 18 %)
Old: 82231 ns, New: 10500 ns (Ratio 13 %)
Old: 37308 ns, New:  6847 ns (Ratio 18 %)
Old: 59884 ns, New:  9384 ns (Ratio 16 %)
Old: 66462 ns, New: 11885 ns (Ratio 18 %)
Old: 55923 ns, New:  8961 ns (Ratio 16 %)
Old: 60616 ns, New:  9500 ns (Ratio 16 %)
Old: 55731 ns, New:  7923 ns (Ratio 14 %)
Old: 25577 ns, New:  5039 ns (Ratio 20 %)
Old: 53923 ns, New:  8346 ns (Ratio 15 %)
Old: 37769 ns, New:  4384 ns (Ratio 12 %)
Old:  7962 ns, New:  2077 ns (Ratio 26 %)
Old: 36808 ns, New:  4039 ns (Ratio 11 %)
Old: 35538 ns, New:  4038 ns (Ratio 11 %)
Old:  9616 ns, New:  2231 ns (Ratio 23 %)
Old: 34269 ns, New:  3885 ns (Ratio 11 %)
Old: 36847 ns, New:  4384 ns (Ratio 12 %)
Old:  7153 ns, New:  1769 ns (Ratio 25 %)
Old: 36615 ns, New:  4269 ns (Ratio 12 %)
```
New implementation is faster, as it takes around 10%-25% of the time of the previous implementation.