@gpesquero 2024-08-25
@rtsisyk: I've just figured out that with the previous implementation with `std::string`, we are able to manage UTF characters for decimal and grouping separators, such as NNBS (NarrowNonBreakingSpace = "\u202F").

But the plain C version that I just tested works Ok with regular chars, but not with UTF chars. We'd have to use `wchar_t`, convert it to `std::wstring` and then to `std::string` to make it work properly, which may complicate the code and we may not get any better performance.

So I'd leave the implementation as proposed, using `std::string` and no low-level C style functions.