@biodranik 2024-08-25
@gpesquero I'm curious, what was your C implementation?