@gpesquero 2024-08-25
> @gpesquero I'm curious, what was your C implementation?

Here's the piece of code: 
https://github.com/gpesquero/organicmaps/commit/e987628b32079840efdb44f456579d8133423a32#diff-0723736751723942b891b3be1579d306bfa25cfa0beaefeb809e77fb7b3e383bR133-R184