@biodranik 2024-08-25
Inserting a substring into a string is tricky in C, and obvious in C++.